Updated Recipes
All recipes are now formatted to have the script.

The ingredient changing script relies on: jQuery and dec2frac.js

Added pictures to the cookie, enchilada, and fajita recipes.

As a note all recipes were initially formatted with printing in mind. If
we need to change the formatting however it shouldn't be hard.